+++ apps/maze/generators/HuntAndKillExample.cpp
@@ -9,26 +9,22 @@ bool HuntAndKillExample::Step(World* w) {
 void HuntAndKillExample::Clear(World* world) {
   visited.clear();
   stack.clear();
-  auto sideOver2 = world->GetSize() / 2;
 
-  for (int i = -sideOver2; i <= sideOver2; i++) {
-    for (int j = -sideOver2; j <= sideOver2; j++) {
+  for (int i = 0; i < world->GetHeight(); i++) {
+    for (int j = 0; j < world->GetWidth(); j++) {
       visited[i][j] = false;
     }
   }
 }
 Point2D HuntAndKillExample::randomStartPoint(World* world) {
   // Todo: improve this if you want
-  auto sideOver2 = world->GetSize() / 2;
-
-  for (int y = -sideOver2; y <= sideOver2; y++)
-    for (int x = -sideOver2; x <= sideOver2; x++)
+  for (int y = 0; y < world->GetHeight(); y++)
+    for (int x = 0; x < world->GetWidth(); x++)
       if (!visited[y][x]) return {x, y};
   return {INT_MAX, INT_MAX};
 }
 
 std::vector<Point2D> HuntAndKillExample::getVisitables(World* w, const Point2D& p) {
-  auto sideOver2 = w->GetSize() / 2;
   std::vector<Point2D> visitables;
 
   // todo: code this
@@ -37,7 +33,6 @@ std::vector<Point2D> HuntAndKillExample::getVisitables(World* w, const Point2D& 
 }
 std::vector<Point2D> HuntAndKillExample::getVisitedNeighbors(World* w, const Point2D& p) {
   std::vector<Point2D> deltas = {{-1, 0}, {0, -1}, {1, 0}, {0, 1}};
-  auto sideOver2 = w->GetSize() / 2;
   std::vector<Point2D> neighbors;
 
   // todo: code this