+++ apps/maze/generators/PrimExample.cpp
@@ -3,8 +3,6 @@
 #include "Random.h"
 
 bool PrimExample::Step(World* w) {
-  int sideOver2 = w->GetSize() / 2;
-
   // todo: code this
 
   return true;
@@ -15,7 +13,6 @@ void PrimExample::Clear(World* world) {
 }
 
 std::vector<Point2D> PrimExample::getVisitables(World* w, const Point2D& p) {
-  auto sideOver2 = w->GetSize() / 2;
   std::vector<Point2D> visitables;
   auto clearColor = Color32(169.0f / 255.0f, 169.0f / 255.0f, 169.0f / 255.0f, 1.0f);  // dark gray
 
@@ -26,7 +23,6 @@ std::vector<Point2D> PrimExample::getVisitables(World* w, const Point2D& p) {
 
 std::vector<Point2D> PrimExample::getVisitedNeighbors(World* w, const Point2D& p) {
   std::vector<Point2D> deltas = {Point2D(0, -1), Point2D(0, 1), Point2D(-1, 0), Point2D(1, 0)};  // N, S, W, E
-  auto sideOver2 = w->GetSize() / 2;
   std::vector<Point2D> neighbors;
 
   // todo: code this